backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-google-genai
langgraph
duckduckgo-search
httpx
pytest 